+++ editor_patch/level.cpp
@@ -619,6 +619,7 @@ CodeInjection CLevelDialog_OnInitDialog_patch{
         std::snprintf(buffer, sizeof(buffer), "%.3f", alpine_level_props.static_mesh_ambient_light_modifier);
         SetDlgItemTextA(hdlg, IDC_MESH_AMBIENT_LIGHT_MODIFIER, buffer);
         CheckDlgButton(hdlg, IDC_RF2_STYLE_GEOMOD, alpine_level_props.rf2_style_geomod ? BST_CHECKED : BST_UNCHECKED);
+        CheckDlgButton(hdlg, IDC_REQUIRES_D3D11, alpine_level_props.requires_d3d11 ? BST_CHECKED : BST_UNCHECKED);
     },
 };
 
@@ -643,6 +644,7 @@ CodeInjection CLevelDialog_OnOK_patch{
             alpine_level_props.static_mesh_ambient_light_modifier = modifier;
         }
         alpine_level_props.rf2_style_geomod = IsDlgButtonChecked(hdlg, IDC_RF2_STYLE_GEOMOD) == BST_CHECKED;
+        alpine_level_props.requires_d3d11 = IsDlgButtonChecked(hdlg, IDC_REQUIRES_D3D11) == BST_CHECKED;
     },
 };